backend/requirements.txt
requests==2.27.1
scikit-learn==1.0.2
numpy==1.21.2
aiofiles==0.8.0
python-multipart==0.0.5
pandas==1.4.1
python-socketio==5.5.2

aiohttp==3.8.1
aiohttp_cors
aiohttp-middlewares